feat : Console Controller와 Spring Controller 분리
