+++ src/main/java/chess/ConsoleChessApplication.java
@@ -1,6 +1,6 @@
 package chess;
 
-import chess.controller.ChessController;
+import chess.controller.console.ChessController;
 
 public class ConsoleChessApplication {
     public static void main(String[] args) {
