+++ src/main/java/chess/controller/spring/ExceptionAdvice.java
@@ -1,0 +1,14 @@
+package chess.controller.spring;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+@ControllerAdvice
+public class ExceptionAdvice {
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<Object> handleAll(Exception e) {
+        return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+}
